Add Slow implementation with removed USize overflows
The earlier Nat version, which the current fast version is based on
was not suitable to 1-1 replacement of Nat with USize, since some of
the Nat-additions in mergeChunksIntoAux could potentially overflow (as
USize-additions) if extremely large arrays were being sorted.

I think all the overflows are gone, assuming at least that chunkSize <
arr.usize, but only time will tell.