Finish pairwise_le_iff_ascending_le_of_trans
Now I can say with confidence that my definition of arrays being in
sorted order is equivalent (in a sense) to the definition of lists
being in sorted order that is found in the Lean standard library. This
will help in the future when I want to show that mergeSort sorts
arrays in the same way that the standard library mergeSort sorts
lists.